+++ app/src/components/billing/BillingContentTabs/PaymentMethods/AddPaymentMethodButton.tsx
@@ -1,29 +1,10 @@
 import { useState } from "react";
-import {
-  Modal,
-  ModalOverlay,
-  ModalContent,
-  ModalHeader,
-  ModalCloseButton,
-  ModalBody,
-  HStack,
-  VStack,
-  Icon,
-  Text,
-  Button,
-  Box,
-  useDisclosure,
-  type UseDisclosureReturn,
-  Spinner,
-} from "@chakra-ui/react";
+import { HStack, Icon, Text, Button, useDisclosure } from "@chakra-ui/react";
 
 import { useHandledAsyncCallback, useSelectedProject } from "~/utils/hooks";
 import { api } from "~/utils/api";
 import { BsPlus } from "react-icons/bs";
-import { Elements, PaymentElement, useElements, useStripe } from "@stripe/react-stripe-js";
-import { loadStripe } from "@stripe/stripe-js";
-import { toast } from "~/theme/ChakraThemeProvider";
-import { env } from "~/env.mjs";
+import PaymentDetailsModal from "./AddPaymentMethodModal";
 
 export default function AddPaymentMethodButton() {
   const disclosure = useDisclosure();
@@ -61,144 +42,5 @@ export default function AddPaymentMethodButton() {
       </Button>
       <PaymentDetailsModal clientSecret={clientSecret} disclosure={disclosure} />
     </>
-  );
-}
-
-// Call `loadStripe` outside of a component’s render to avoid recreating the `Stripe` object on every render.
-const stripePromise = loadStripe(env.NEXT_PUBLIC_STRIPE_PUBLISHABLE_KEY ?? "");
-
-const PaymentDetailsModal = ({
-  clientSecret,
-  disclosure,
-}: {
-  clientSecret: string;
-  disclosure: UseDisclosureReturn;
-}) => {
-  const options = {
-    clientSecret,
-    // Fully customizable with appearance API.
-    appearance: {
-      /*...*/
-    },
-  };
-
-  return (
-    <Modal
-      size={{ base: "xl", md: "2xl" }}
-      closeOnOverlayClick={false}
-      closeOnEsc={false}
-      {...disclosure}
-    >
-      <ModalOverlay />
-      <ModalContent maxW="95%" h="auto">
-        <ModalHeader>
-          <HStack>
-            <Text>Add payment method</Text>
-          </HStack>
-        </ModalHeader>
-        <ModalCloseButton />
-        <ModalBody overflowY="auto" p={8}>
-          <Box w="full">
-            <VStack w="full" justifyContent={"center"}>
-              <Elements stripe={stripePromise} options={options}>
-                <SetupStripePaymentMethodForm disclosure={disclosure} clientSecret={clientSecret} />
-              </Elements>
-            </VStack>
-          </Box>
-        </ModalBody>
-      </ModalContent>
-    </Modal>
-  );
-};
-
-function SetupStripePaymentMethodForm({
-  disclosure,
-  clientSecret,
-}: {
-  disclosure: UseDisclosureReturn;
-  clientSecret: string;
-}) {
-  const stripe = useStripe();
-  const elements = useElements();
-  const selectedProject = useSelectedProject().data;
-  const utils = api.useContext();
-
-  // Code from stripe docs
-  const [handleAddPaymentDetail, addPaymentDetailLoading] = useHandledAsyncCallback(async () => {
-    const setupIntent = await addPaymentDetail();
-
-    if (setupIntent) {
-      switch (setupIntent.status) {
-        case "succeeded":
-          toast({
-            description: "Success! Your payment method has been saved.",
-            status: "success",
-            duration: 5000,
-            isClosable: true,
-          });
-          break;
-
-        case "processing":
-          toast({
-            description: "Processing payment details. Please wait a few minutes.",
-            status: "success",
-            duration: 5000,
-            isClosable: true,
-          });
-          break;
-
-        case "requires_payment_method":
-          toast({
-            description: "Failed to process payment details. Please try another payment method.",
-            status: "error",
-            duration: 5000,
-            isClosable: true,
-          });
-          break;
-      }
-      disclosure.onClose();
-    }
-
-    await utils.payments.getPaymentMethods.invalidate();
-  }, [utils, addPaymentDetail]);
-
-  async function addPaymentDetail() {
-    if (!stripe || !elements) {
-      return;
-    }
-
-    const { error } = await stripe.confirmSetup({
-      elements,
-      confirmParams: {
-        return_url: `${env.NEXT_PUBLIC_HOST}/p/${
-          selectedProject?.slug ?? ""
-        }/billing/payment-methods`,
-      },
-      redirect: "if_required",
-    });
-
-    if (error) {
-      toast({
-        description: error.message,
-        status: "error",
-        duration: 5000,
-        isClosable: true,
-      });
-    } else {
-      const { setupIntent } = await stripe.retrieveSetupIntent(clientSecret);
-
-      return setupIntent;
-    }
-  }
-
-  return (
-    <Box style={{ width: "100%" }}>
-      <PaymentElement />
-      <HStack w="full" justifyContent="end">
-        <Button colorScheme="gray" minW={24} onClick={() => handleAddPaymentDetail()}>
-          {addPaymentDetailLoading ? <Spinner /> : "Save"}
-        </Button>
-      </HStack>
-    </Box>
   );
 }
